add hooks as an option
